csharp/structs review points

+++ languages/csharp/exercises/concept/structs/Structs.cs
@@ -8,9 +8,8 @@ public struct Coord
         Y = y;
     }
 
-    public int X {get; }
+    public ushort X {get; }
     public ushort Y {get; }
-
 }
 
 public struct Plot
@@ -41,4 +40,3 @@ public class ClaimsHandler
         throw new NotImplementedException("Please implement the ClaimsHandler.GetClaimWithLongestSide() method");
     }
 }
-
